修复表达式求值系统的两个关键问题：
1. 运算符优先级一致性修复
   - gen-expr.c: 为生成的表达式字符串添加括号，确保字符串形式与AST计算顺序一致
   - 之前: `sprintf(eval_buf, "%s %c %s", l_eval, op, r_eval)`
   - 现在: `sprintf(eval_buf, "(%s %c %s)", l_eval, op, r_eval)`

2. 新增批量测试框架
   - monitor.c: 添加 `-e/--expr` 参数支持文件输入
   - expr.c: 实现 `test_expr()` 函数，支持"预期结果 表达式"格式的自动化测试
   - 增强调试能力，支持大规模表达式测试

3. 基础设施改进
   - 修正 p 命令的输出显示格式
   - 在 native.mk 中添加 Makefile 中的测试示例配置

+++ nemu/src/monitor/sdb/sdb.c
@@ -206,7 +206,7 @@ static int cmd_p(char *args) {
     return 0;
   }
 
-  printf("Result: " FMT_WORD "(%d)\n", result, (int)result);
+  printf("Result: " FMT_WORD "(%u)\n", result, result);
   return 0;
 }
